heredoc中にctrl+Dが押された場合に改行を入れました。Permission Deny127, syntax error2など返り値もbash本家に合わせて修正しました
