+++ srcs/executor/executor_external.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   executor_external.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: myokono <myokono@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*   By: ryabuki <ryabuki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/07 19:48:16 by myokono           #+#    #+#             */
-/*   Updated: 2025/04/18 19:58:07 by myokono          ###   ########.fr       */
+/*   Updated: 2025/04/21 13:41:27 by ryabuki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,10 @@ static int	handle_execve_error(char *exec_path)
 	if (stat(exec_path, &st) == 0 && S_ISDIR(st.st_mode))
 		print_error(exec_path, "Is a directory");
 	else if (err == EACCES)
+	{
 		print_error(exec_path, "Permission denied");
+		code = 127;
+	}
 	else if (err == ENOENT)
 	{
 		print_error(exec_path, "No such file or directory");
