+++ srcs/gnl/srcs/get_next_line.c
@@ -6,7 +6,7 @@
 /*   By: ryabuki <ryabuki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/19 19:05:04 by yabukirento       #+#    #+#             */
-/*   Updated: 2025/04/17 14:02:16 by ryabuki          ###   ########.fr       */
+/*   Updated: 2025/04/21 13:54:56 by ryabuki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,10 @@ static char	*read_to_newline(int fd, char **saved, char *buffer)
 			break ;
 	}
 	if ((bytes_read == 0 && ft_strlen(*saved) == 0) || bytes_read < 0)
+	{
+		write(STDOUT_FILENO, "\n", 1);
 		return (free(*saved), NULL);
+	}
 	return (*saved);
 }
