+++ srcs/parser/lexer.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   lexer.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yabukirento <yabukirento@student.42.fr>    +#+  +:+       +#+        */
+/*   By: ryabuki <ryabuki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/27 12:35:52 by myokono           #+#    #+#             */
-/*   Updated: 2025/04/16 21:50:13 by yabukirento      ###   ########.fr       */
+/*   Updated: 2025/04/21 13:28:12 by ryabuki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
